update coeff, making it static
